requirements.txt
pyautogui==0.9.54
pyperclip==1.8.2
Pillow==10.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
supabase==2.1.0 
